% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bibliography}
\alias{bibliography}
\title{Data bibliography.}
\format{A data frame with six variables.
 \describe{
     \item{\code{obs}}{ID}
     \item{\code{id}}{Gutenberg ID}
     \item{\code{utf8}}{Indicator of utf8 encoding}
     \item{\code{title}}{Raw title}
     \item{\code{gTitle}}{Gutenberg title}
     \item{\code{year_comp}}{Year completed}
     \item{\code{last_word}}{last word}
 }}
\usage{
bibliography
}
\description{
Type of text encoding used for Dickens's works. Used to determine file path for download.
}
\keyword{datasets}
