% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dickens.utf}
\alias{dickens.utf}
\title{Gutenberg text file encoding.}
\format{A data frame with two variables.
 \describe{
     \item{\code{ID}}{Gutenberg ID}
     \item{\code{utf8}}{Indicator of utf8 encoding}
 }}
\usage{
dickens.utf
}
\description{
Type of text encoding used for Dickens's works. Used to determine file path for download.
}
\keyword{datasets}
