% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWork.R
\name{getWork}
\alias{getWork}
\title{Download work from gutenberg.org}
\usage{
getWork(id, g_title, last_text)
}
\arguments{
\item{id}{Numeric. Gutenberg ID of work.}

\item{g_title}{Character. Start of text (gutenberg title).}

\item{last_text}{Character. End of text (last word, include punctuation).}
}
\value{
quanteda object.
}
\description{
Gets txt file.
}
