% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWork2.R
\name{getWork2}
\alias{getWork2}
\title{Download work from gutenberg.org}
\usage{
getWork2(id, g_title, last_text, utf_8 = TRUE)
}
\arguments{
\item{id}{Numeric. Gutenberg ID of work.}

\item{g_title}{Character. Start of text (gutenberg title).}

\item{last_text}{Character. End of text (last word, include punctuation).}

\item{utf_8}{Logical. Flag for url.}
}
\value{
quanteda object.
}
\description{
Gets txt file.
}
