% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bibliography}
\alias{bibliography}
\title{Data bibliography.}
\format{A data frame with six variables.
 \describe{
     \item{\code{obs}}{ID}
     \item{\code{ID}}{Gutenberg ID}
     \item{\code{utf8}}{Indicator of utf8 encoding}
     \item{\code{Title}}{Raw title}
     \item{\code{Title2}}{Clean title}
     \item{\code{Year_Comp}}{Year completed}
 }}
\usage{
bibliography
}
\description{
Type of text encoding used for Dickens's works. Used to determine file path for download.
}
\keyword{datasets}
