% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWork.R
\name{getWork}
\alias{getWork}
\title{Get Work.}
\usage{
getWork(id, first, last)
}
\arguments{
\item{id}{Numeric. Gutenberg ID.}

\item{first}{Character. First word.}

\item{last}{Character. Last word.}
}
\value{
quanteda object
}
\description{
Fetch work" from gutenberg.org for 'qunateda'.
}
\examples{
# A Christmas Carol:
getWork(46, "PREFACE", "One")
}
