% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWork.R
\name{getWork}
\alias{getWork}
\title{Get Work.}
\usage{
getWork(id)
}
\arguments{
\item{id}{Numeric. Gutenberg ID.}
}
\value{
quanteda object
}
\description{
Fetch work" from gutenberg.org for 'qunateda'.
}
\examples{
# A Christmas Carol:
getWork(46)

# A Tale of Two Cities:
getWork(98)
}
